+++ lib/src/paged_datatable.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:calendar_date_picker2/calendar_date_picker2.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -30,26 +31,62 @@ part 'types.dart';
 /// A paginated DataTable that allows page caching and filtering
 /// [TKey] is the type of the page token
 /// [TResult] is the type of data the data table will show.
-class PagedDataTable<TKey extends Object, TResult extends Object>
+class PagedDataTable<TKey extends Comparable, TResultId extends Comparable, TResult extends Object>
     extends StatelessWidget {
+  /// The callback that gets executed when a page is fetched.
   final FetchCallback<TKey, TResult> fetchPage;
+
+  /// The initial page to fetch.
   final TKey initialPage;
+
+  /// The list of filters to show.
   final List<TableFilter>? filters;
-  final PagedDataTableController<TKey, TResult>? controller;
+
+  /// A custom controller used to programatically control the table.
+  final PagedDataTableController<TKey, TResultId, TResult>? controller;
+
+  /// The list of columns to display.
   final List<BaseTableColumn<TResult>> columns;
+
+  /// A custom menu tooltip to show in the filter bar.
   final PagedDataTableFilterBarMenu? menu;
-  final Widget? footer, header;
+
+  /// A custom widget to build in the footer, aligned to the left.
+  ///
+  /// Navigation widgets remain untouched.
+  final Widget? footer;
+
+  /// A custom widget to build in the footer, aligned to the left.
+  ///
+  /// Filter widgets remain untouched.
+  final Widget? header;
+
+  /// A custom builder that display any error.
   final ErrorBuilder? errorBuilder;
+
+  /// A custom builder that builds when no item is found.
   final WidgetBuilder? noItemsFoundBuilder;
+
+  /// A custom theme to apply only to this DataTable instance.
   final PagedDataTableThemeData? theme;
+
+  /// Indicates if the table allows the user to select rows.
   final bool rowsSelectable;
+
+  /// A custom builder that builds a row.
   final CustomRowBuilder<TResult>? customRowBuilder;
+
+  /// A stream to listen and refresh the table when any update is received.
   final Stream? refreshListener;
+
+  /// A function that returns the id of an item.
+  final ModelIdGetter<TResultId, TResult> idGetter;
 
   const PagedDataTable(
       {required this.fetchPage,
       required this.initialPage,
       required this.columns,
+      required this.idGetter,
       this.filters,
       this.menu,
       this.controller,
@@ -65,20 +102,20 @@ class PagedDataTable<TKey extends Object, TResult extends Object>
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<_PagedDataTableState<TKey, TResult>>(
+    return ChangeNotifierProvider<_PagedDataTableState<TKey, TResultId, TResult>>(
       create: (context) => _PagedDataTableState(
           columns: columns,
           rowsSelectable: rowsSelectable,
           filters: filters,
+          idGetter: idGetter,
           controller: controller,
           fetchCallback: fetchPage,
           initialPage: initialPage,
           refreshListener: refreshListener),
       builder: (context, widget) {
-        var state = context.read<_PagedDataTableState<TKey, TResult>>();
-        final localTheme = PagedDataTableTheme.maybeOf(context) ??
-            theme ??
-            _kDefaultPagedDataTableTheme;
+        var state = context.read<_PagedDataTableState<TKey, TResultId, TResult>>();
+        final localTheme =
+            PagedDataTableTheme.maybeOf(context) ?? theme ?? _kDefaultPagedDataTableTheme;
 
         Widget child = Material(
           color: localTheme.backgroundColor,
@@ -86,36 +123,38 @@ class PagedDataTable<TKey extends Object, TResult extends Object>
           textStyle: localTheme.textStyle,
           shape: theme?.border,
           child: LayoutBuilder(builder: (context, constraints) {
-            var width = constraints.maxWidth - (columns.length * 32);
+            var width = constraints.maxWidth - (columns.length * 32) - (rowsSelectable ? 32 : 0);
             state.availableWidth = width;
             return Column(
               children: [
                 /* FILTER TAB */
-                if (header != null ||
-                    menu != null ||
-                    state.filters.isNotEmpty) ...[
-                  _PagedDataTableFilterTab<TKey, TResult>(menu, header),
+                if (header != null || menu != null || state.filters.isNotEmpty) ...[
+                  _PagedDataTableFilterTab<TKey, TResultId, TResult>(menu, header),
                   Divider(height: 0, color: localTheme.dividerColor),
                 ],
 
                 /* HEADER ROW */
-                _PagedDataTableHeaderRow<TKey, TResult>(rowsSelectable, width),
+                _PagedDataTableHeaderRow<TKey, TResultId, TResult>(rowsSelectable, width),
                 Divider(height: 0, color: localTheme.dividerColor),
 
                 /* ITEMS */
                 Expanded(
-                  child: _PagedDataTableRows<TKey, TResult>(
+                  child: _PagedDataTableRows<TKey, TResultId, TResult>(
                       rowsSelectable,
-                      customRowBuilder,
+                      customRowBuilder ??
+                          CustomRowBuilder<TResult>(
+                              builder: (context, item) =>
+                                  throw UnimplementedError("This does not build nothing"),
+                              shouldUse: (context, item) => false),
                       noItemsFoundBuilder,
                       errorBuilder,
                       width),
                 ),
 
                 /* FOOTER */
-                if (theme?.configuration.footer.visible ?? true) ...[
+                if (localTheme.configuration.footer.footerVisible) ...[
                   Divider(height: 0, color: localTheme.dividerColor),
-                  _PagedDataTableFooter<TKey, TResult>(footer)
+                  _PagedDataTableFooter<TKey, TResultId, TResult>(footer)
                 ]
               ],
             );