+++ lib/src/paged_datatable_controller.dart
@@ -3,23 +3,39 @@
 part of 'paged_datatable.dart';
 
 /// Represents a controller of a [PagedDataTable]
-class PagedDataTableController<TKey extends Object, TResult extends Object> {
-  late final _PagedDataTableState<TKey, TResult> _state;
+class PagedDataTableController<TKey extends Comparable, TResultId extends Comparable,
+    TResult extends Object> {
+  late final _PagedDataTableState<TKey, TResultId, TResult> _state;
 
-  /// Returns the current showing dataset elements.
-  Iterable<TResult> get currentDataset => _state._rowsState.map((e) => e.item);
+  /// Returns the current showing dataset elements as an unmodifiable list.
+  List<TResult> get currentDataset => UnmodifiableListView(_state._items);
 
-  void dispose() {
-    _state.dispose();
-  }
+  /// Returns the actual pagination info
+  PagedDataTablePaginationInfo get paginationInfo => PagedDataTablePaginationInfo._(
+      currentPage: _state.currentPage,
+      currentPageSize: _state._items.length,
+      hasNextPage: _state.hasNextPage,
+      rowsPerPage: _state._pageSize,
+      hasPreviousPage: _state.hasPreviousPage);
 
-  /// Refreshes the table clearing the cache and fetching from source again.
+  /// Refreshes the table fetching from source again.
   /// If [currentDataset] is true, it will only refresh the current viewing resultset, otherwise,
-  /// it will clear the local cache and start from page 1.
-  void refresh({bool currentDataset = true}) {
-    _state._refresh(currentDataset: currentDataset);
-    // _controller.add(_ControllerEvent(_ControllerEventType.refresh, null));
-  }
+  /// it will start from page 1.
+  ///
+  /// The future completes when the fetch is done.
+  Future<void> refresh({bool currentDataset = true}) => _state._refresh(initial: !currentDataset);
+
+  /// Advances to the next page.
+  ///
+  /// If there is no next page, this method will fail.
+  /// The future completes when the fetch is done.
+  Future<void> advancePage() => _state.nextPage();
+
+  /// Backs off to the previous page.
+  ///
+  /// If there is no previous page, this method will fail.
+  /// The future completes when the fetch is done.
+  Future<void> backPage() => _state.previousPage();
 
   /// Sets a filter and fetches items from source.
   void setFilter(String id, dynamic value) {
@@ -46,42 +62,22 @@ class PagedDataTableController<TKey extends Object, TResult extends Object> {
     return _state.filters[filterName]?.value;
   }
 
-  /// Gets all the selected rows in the current resultset
+  /// Returns a list of the selected items.
   List<TResult> getSelectedRows() {
-    return _state.selectedRows.entries
-        .where((element) => element.value)
-        .map((e) => _state.tableCache.currentResultset[e.key])
-        .toList();
+    return _state.selectedRows.values.map((e) => _state._items[e]).toList(growable: false);
   }
 
   /// Unselects any selected row in the current resultset
-  void unselectAllRows() {
-    _state.selectedRows.clear();
-    for (var rowState in _state._rowsState) {
-      if (rowState._isSelected) {
-        rowState.selected = false;
-      }
-    }
-  }
+  void unselectAllRows() => _state.unselectAllRows();
 
-  /// Marks the row at [index] as selected
-  void selectRow(int index) {
-    _state.selectedRows[index] = true;
-    _state._rowsState[index].selected = true;
-  }
+  /// Selects all the rows in the current resultset
+  void selectAllRows() => _state.selectAllRows();
 
-  /// Updates an item from the current resultset located at [rowIndex] and rebuilds the row.
-  void modifyRowValue(int rowIndex, void Function(TResult item) update) {
-    try {
-      var row = _state.tableCache.currentResultset[rowIndex];
-      update(row);
+  /// Marks the row whose id is [itemId] as unselected
+  void unselectRow(TResultId itemId) => _state.unselectRow(itemId);
 
-      // refresh state of that row.
-      _state._rowsState[rowIndex].refresh();
-    } catch (_) {
-      throw TableError("There is no row at index $rowIndex.");
-    }
-  }
+  /// Marks the row whose id is [itemId] as selected
+  void selectRow(TResultId itemId) => _state.selectRow(itemId);
 
   /// Builds every row that matches [predicate].
   void refreshRowWhere(bool Function(TResult element) predicate) {
@@ -92,40 +88,77 @@ class PagedDataTableController<TKey extends Object, TResult extends Object> {
   }
 
   /// Updates every item from the current resultset that matches [predicate] and rebuilds it.
-  void modifyRowsValue(bool Function(TResult element) predicate,
-      void Function(TResult item) update) {
-    var matched = _state.tableCache.currentResultset.where(predicate);
-    for (var match in matched) {
-      update(match);
+  ///
+  /// Keep in mind this method will iterate over every item in the current dataset, so, if the dataset is large, it can be costly.
+  void modifyRowsValue(
+      bool Function(TResult element) predicate, void Function(TResult item) update) {
+    int index = 0;
+    for (final item in _state._items) {
+      if (predicate(item)) {
+        update(item);
+        _state._rowsState[index].refresh();
+      }
 
-      // refresh state of that row.
-      _state._rowsState
-          .firstWhere((element) => predicate(element.item))
-          .refresh();
+      index++;
     }
   }
 
-  /// Rebuilds the row at the specified [rowIndex] to reflect changes to the item.
-  void refreshRow(int rowIndex) {
-    try {
-      // refresh state of that row.
-      _state._rowsState[rowIndex].refresh();
-    } catch (_) {
-      throw TableError("There is no row at index $rowIndex.");
+  /// Updates an item from the current resultset with the id [itemId] and rebuilds the row.
+  void modifyRowValue(TResultId itemId, void Function(TResult item) update) {
+    final rowIndex = _state._rowsStateMapper[itemId];
+    if (rowIndex == null) {
+      throw TableError('Item with key "$itemId" is not in the current dataset.');
     }
+
+    final row = _state._rowsState[rowIndex];
+    final item = _state._items[row.index];
+    update(item);
+
+    // refresh state of that row.
+    row.refresh();
   }
 
-  /// Removes the row containing [element].
-  /// Keep in mind this will work only if [TResult] defines a custom hashcode implementation.
-  void removeRow(TResult element) {
-    var rowStateIndex =
-        _state._rowsState.indexWhere((elem) => elem.item == element);
-    if (rowStateIndex != -1) {
-      _state.tableCache.deleteFromCurrentDataset(element);
-      _state._rowsState.removeAt(rowStateIndex);
-      _state._rowsChange = rowStateIndex;
-      // ignore: invalid_use_of_visible_for_testing_member
-      _state.notifyListeners();
+  /// Rebuilds the row which has the specified [itemId] to reflect changes to the item.
+  void refreshRow(TResultId itemId) {
+    final rowIndex = _state._rowsStateMapper[itemId];
+    if (rowIndex == null) {
+      throw TableError('Item with key "$itemId" is not in the current dataset.');
     }
+
+    _state._rowsState[rowIndex].refresh();
   }
+
+  /// Removes the row containing an element whose id is [itemId]
+  void removeRow(TResultId itemId) {
+    final rowIndex = _state._rowsStateMapper[itemId];
+    if (rowIndex == null) {
+      throw TableError('Item with key "$itemId" is not in the current dataset.');
+    }
+
+    _state._rowsState.removeAt(rowIndex);
+    _state._items.removeAt(rowIndex);
+    _state._rowsStateMapper.remove(itemId);
+    _state._rowsChange = rowIndex;
+    // ignore: invalid_use_of_visible_for_testing_member
+    _state.notifyListeners();
+  }
+
+  /// Disposes the controller.
+  ///
+  /// After this method is called, the DataTable is disposed and cannot be used.
+  void dispose() {
+    _state.dispose();
+  }
+}
+
+class PagedDataTablePaginationInfo {
+  final bool hasNextPage, hasPreviousPage;
+  final int currentPageSize, currentPage, rowsPerPage;
+
+  PagedDataTablePaginationInfo._(
+      {required this.hasNextPage,
+      required this.hasPreviousPage,
+      required this.currentPageSize,
+      required this.currentPage,
+      required this.rowsPerPage});
 }