+++ lib/src/paged_datatable_row_state.dart
@@ -1,15 +1,18 @@
 part of 'paged_datatable.dart';
 
-class _PagedDataTableRowState<TResult extends Object> extends ChangeNotifier {
+class _PagedDataTableRowState<TResultId extends Comparable, TResult extends Object>
+    extends ChangeNotifier {
   final TResult item;
-  final int rowIndex;
+  final TResultId itemId;
+  final int index;
+
   bool _isSelected = false;
   set selected(bool newValue) {
     _isSelected = newValue;
     notifyListeners();
   }
 
-  _PagedDataTableRowState(this.item, this.rowIndex);
+  _PagedDataTableRowState(this.index, this.item, this.itemId);
 
   void refresh() {
     notifyListeners();